+++ src/main/java/org/octri/fhir_sandbox/service/SampleDataService.java
@@ -2,6 +2,7 @@ package org.octri.fhir_sandbox.service;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
+import java.time.Duration;
 import java.util.Arrays;
 
 import org.apache.commons.logging.Log;
@@ -10,6 +11,8 @@ import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.Resource;
 import org.octri.fhir_sandbox.config.FhirServerProperties;
 import org.octri.fhir_sandbox.config.SandboxDataConfig;
+import org.octri.fhir_sandbox.hapi.RenewableBearerTokenInterceptor;
+import org.octri.fhir_sandbox.oauth2.utils.OAuthUtils;
 import org.octri.fhir_sandbox.util.FhirDataUtil;
 import org.springframework.core.io.support.ResourcePatternResolver;
 import org.springframework.stereotype.Service;
@@ -25,25 +28,29 @@ import ca.uhn.fhir.rest.client.api.IGenericClient;
 public class SampleDataService {
 
 	private static final Log log = LogFactory.getLog(SampleDataService.class);
+	private static final Duration TOKEN_DURATION = Duration.ofMinutes(5);
 
 	private final SandboxDataConfig dataConfig;
 	private final ResourcePatternResolver resourcePatternResolver;
 	private final FhirServerProperties fhirServerProperties;
 	private final FhirContext fhirContext;
+	private final PreAuthorizedTokenService tokenService;
 
 	public SampleDataService(SandboxDataConfig dataConfig, ResourcePatternResolver resourcePatternResolver,
-			FhirServerProperties fhirServerProperties, FhirContext fhirContext) {
+			FhirServerProperties fhirServerProperties, FhirContext fhirContext,
+			PreAuthorizedTokenService tokenService) {
 		this.dataConfig = dataConfig;
 		this.resourcePatternResolver = resourcePatternResolver;
 		this.fhirServerProperties = fhirServerProperties;
 		this.fhirContext = fhirContext;
+		this.tokenService = tokenService;
 
 		fhirContext.getRestfulClientFactory().setSocketTimeout(this.fhirServerProperties.getSocketTimeout());
 	}
 
 	/**
 	 * Uses the provided pattern to resolve {@link Resource} objects, then parses them using the {@link Bundle} model.
-	 * 
+	 *
 	 * @param resourcePattern
 	 * @return
 	 * @throws UncheckedIOException
@@ -62,7 +69,7 @@ public class SampleDataService {
 
 	/**
 	 * Executes the client transaction to save the {@link Bundle} to the remote server.
-	 * 
+	 *
 	 * @param bundle
 	 * @param client
 	 */
@@ -76,19 +83,26 @@ public class SampleDataService {
 
 	/**
 	 * Reads configured sample data then posts them to the FHIR server.
-	 * 
-	 * Intentially does not handle errors from helper methods.
-	 * 
+	 *
+	 * Intentionally does not handle errors from helper methods.
+	 *
 	 * @param fhirUrl
 	 * @throws UncheckedIOException
 	 * @throws DataFormatException
 	 */
 	public void loadSampleData(String fhirUrl) {
 		var fhirClient = fhirContext.newRestfulGenericClient(fhirUrl);
+		fhirClient.registerInterceptor(createTokenInterceptor(fhirUrl));
 		for (var pattern : dataConfig.getSampleResourcePatterns()) {
 			for (var bundle : getSampleBundles(pattern)) {
 				handleTransaction(bundle, fhirClient);
 			}
 		}
 	}
+
+	private RenewableBearerTokenInterceptor createTokenInterceptor(String fhirUrl) {
+		return new RenewableBearerTokenInterceptor(tokenService, OAuthUtils.getSystemClaimsForSandbox(fhirUrl),
+				TOKEN_DURATION);
+	}
+
 }